debundle: rename logical_modules.rs -> lowering/mod.rs
Mechanical rename in prep for splitting the 4500-line file into
per-concern siblings under lowering/. No content change in this commit;
just the file move plus BUILD.bazel target/srcs rename and the one
pipeline.rs import (use logical_modules::... -> use lowering::...).

Why "lowering": the file implements the lowering pipeline that
materializes a parsed chunk + spec into emitted modules. The old name
"logical_modules" names the spec-side concept (one chunk decomposes
into N logical modules) but the file's contents are the implementation
of the lowering process, not the data type itself.

BAZEL_TEST_INVOCATIONS=none: file rename + build target rename, no
semantic change.
